feat: replace `filepicker` with `zoxide`
Before:
in the `new session` screen, ctrl+f would send a msg to `filepicker`
(alias for zellij's strider plugin) which would launch it in a floating
pane for the user to pick a folder.

After:

same functionality but replace `filepicker` integration with zoxide.
low-level details:

use `zellij action new-pane --close-on-exit` to be able to close the
command after it's done, capture its output via `zellij pipe`.
Current plugin commands don't allow to launch a floating pane, that
closes itself after the cmd is done AND capture its output.